--- tests/map/scenes/test_ascii.py
@@ -1,16 +1,0 @@
-from metta.map.scenes.ascii import Ascii
-from tests.map.scenes.utils import assert_grid, render_scene
-
-
-def test_basic():
-    scene = render_scene(Ascii.factory(Ascii.Params(uri="tests/map/scenes/fixtures/test.map")), (4, 4))
-
-    assert_grid(
-        scene,
-        """
-####
-#_.#
-##.#
-####
-        """,
-    )
