+++ tests/map/scenes/test_random.py
@@ -1,0 +1,28 @@
+from metta.map.scenes.random import Random
+from tests.map.scenes.utils import render_scene
+
+
+def test_objects():
+    scene = render_scene(
+        Random.factory(Random.Params(objects={"altar": 3, "temple": 2})),
+        (3, 3),
+    )
+
+    assert (scene.grid == "altar").sum() == 3
+    assert (scene.grid == "temple").sum() == 2
+
+
+def test_agents():
+    scene = render_scene(Random.factory(Random.Params(agents=2)), (3, 3))
+
+    assert (scene.grid == "agent.agent").sum() == 2
+
+
+def test_agents_dict():
+    scene = render_scene(
+        Random.factory(Random.Params(agents={"prey": 2, "predator": 1})),
+        (3, 3),
+    )
+
+    assert (scene.grid == "agent.prey").sum() == 2
+    assert (scene.grid == "agent.predator").sum() == 1
